pyproject.toml
[tool.poetry]
name = "chart"
version = "0.1.0"
description = ""
authors = ["Gilad Kutiel <gilad.kutiel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
matplotlib = "^3.9.2"
python-bidi = "^0.6.3"
arabic-reshaper = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
